replaced the tracer solution with a variable-based solution
