+++ packages/message-parser/loaders/pegtransform.js
@@ -5,8 +5,6 @@ module.exports = {
     code: pegjs.generate(content, {
       output: 'source',
       format: 'commonjs',
-      // We rely on a tracer to prevent recurrence on rules like Bold, Italic, Strikethrough and References
-      trace: true,
     }),
   }),
 };
