+++ packages/message-parser/src/index.ts
@@ -1,22 +1,26 @@
 import type { Root } from './definitions';
 import * as grammar from './grammar.pegjs';
-import { type Options, Tracer } from './tracer';
 
 export * from './definitions';
 
 export { isNodeOfType } from './guards';
 
-export const parse = (input: string, options?: Options): Root => {
-  const parserOptions = Tracer.createParserOptions(options);
-
-  const result = grammar.parse(input, parserOptions);
-  return result;
+export type Options = {
+  colors?: boolean;
+  emoticons?: boolean;
+  katex?: {
+    dollarSyntax?: boolean;
+    parenthesisSyntax?: boolean;
+  };
+  customDomains?: string[];
 };
+
+export const parse = (input: string, options?: Options): Root =>
+  grammar.parse(input, options);
 
 export {
   /** @deprecated */
   parse as parser,
   /** @deprecated */
   Root as MarkdownAST,
-  Options,
 };
