+++ packages/message-parser/src/utils.ts
@@ -198,22 +198,19 @@ const joinEmoji = (
 export const reducePlainTexts = (
   values: Paragraph['value']
 ): Paragraph['value'] =>
-  values
-    .flatMap((item) => item)
-    .filter((item) => item)
-    .reduce((result, item, index, values) => {
-      const next = values[index + 1];
-      const current = joinEmoji(item, values[index - 1], next);
-      const previous: Inlines = result[result.length - 1];
+  values.flat().reduce((result, item, index, values) => {
+    const next = values[index + 1];
+    const current = joinEmoji(item, values[index - 1], next);
+    const previous: Inlines = result[result.length - 1];
 
-      if (previous) {
-        if (current.type === 'PLAIN_TEXT' && current.type === previous.type) {
-          previous.value += current.value;
-          return result;
-        }
+    if (previous) {
+      if (current.type === 'PLAIN_TEXT' && current.type === previous.type) {
+        previous.value += current.value;
+        return result;
       }
-      return [...result, current];
-    }, [] as Paragraph['value']);
+    }
+    return [...result, current];
+  }, [] as Paragraph['value']);
 export const lineBreak = (): LineBreak => ({
   type: 'LINE_BREAK',
   value: undefined,
@@ -249,4 +246,14 @@ export const timestamp = (
     },
     fallback: plain(`<t:${value}:${type || 't'}>`),
   };
+};
+
+export const extractFirstResult = (
+  value: Types[keyof Types]['value']
+): Types[keyof Types]['value'] => {
+  if (typeof value !== 'object' || !Array.isArray(value)) {
+    return value;
+  }
+
+  return value.filter((item) => item).shift() as Types[keyof Types]['value'];
 };
